backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-core
pydantic
faiss-cpu
